refactor(web): split navigation and namespace plugin IDs
Separate the first-party catalog, surface policy, plugin mapping, resolver,
and shared types so navigation can evolve without one monolithic module.

Carry plugin ownership through nav registrations and encode owner and item IDs
in destination keys while preserving raw item IDs for compatibility selectors.
