+++ apps/web/components/integrations/integrations-menu.tsx
@@ -23,7 +23,7 @@ const HOVER_CLOSE_DELAY_MS = 180;
 
 /**
  * Configured integration destinations, in manifest order. Ids, labels, hrefs and
- * icons come from `lib/navigation/destinations.ts`; availability comes from
+ * icons come from `lib/navigation/core-destinations.ts`; availability comes from
  * `useIntegrationAvailability`.
  */
 function useIntegrationDestinations() {
