+++ apps/web/components/navigation/destination-rows.test.tsx
@@ -1,7 +1,7 @@
 import { cleanup, fireEvent, render, screen } from "@testing-library/react";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import { DestinationRows } from "./destination-rows";
-import type { ResolvedDestination } from "@/lib/navigation/destinations";
+import type { ResolvedDestination } from "@/lib/navigation/types";
 
 vi.mock("@/lib/routing/client-router", () => ({
   usePathname: () => "/",
